Move all UI related files in their own package
